fixing the view_start and end message. adding a hull message to turn on/off the calculation of the distance to hull. Now the calculation of the hull is done after setting the weights of source. THERE IS NO HULL FOR WEIGHT PRESETS : THE HULL IS COMPUTED FOR ALL DESTINATIONS.
git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@4632 464221e2-0a0f-0410-983b-f8d57d210c17

+++ SourceCode/MaxObjects/jcom.dbap/hull2D.h
@@ -35,13 +35,13 @@ not removed.
 /** Data structures to store a 2 dimensional convex hull */
 
 /*--------- Point(s) Structure ---------*/
-typedef float t_xy[2];	/* Type float point */
+typedef double t_xy[2];	/* Type float point */
 
 typedef struct _structPoint t_structPoint;
 typedef t_structPoint *t_point;
 struct _structPoint {
 	t_xy	v;
-	int		vnum;
+	long	vnum;
 	bool	del;
 	t_xy	p0;
 };
